backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Logging
structlog==23.2.0

# Email & 2FA
aiosmtplib==3.0.1
pyotp==2.9.0
qrcode[pil]==7.4.2

# Analytics (optional)
pandas==2.1.3
polars==0.19.19

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
